split Combustion logic from Step.hs, create separate Util module
