+++ src/Swarm/Game/Step.hs
@@ -30,8 +30,7 @@ import Control.Effect.Error
 import Control.Effect.Lens
 import Control.Effect.Lift
 import Control.Lens as Lens hiding (Const, distrib, from, parts, use, uses, view, (%=), (+=), (.=), (<+=), (<>=))
-import Control.Monad (foldM, forM, forM_, guard, join, msum, unless, when, zipWithM)
-import Data.Array (bounds, (!))
+import Control.Monad (foldM, forM, forM_, guard, msum, unless, when, zipWithM)
 import Data.Bifunctor (second)
 import Data.Bool (bool)
 import Data.Char (chr, ord)
@@ -76,6 +75,8 @@ import Swarm.Game.Scenario.Objective.WinCheck qualified as WC
 import Swarm.Game.Scenario.Topography.Navigation.Portal (Navigation (..), destination, reorientation)
 import Swarm.Game.Scenario.Topography.Navigation.Waypoint (WaypointName (..))
 import Swarm.Game.State
+import Swarm.Game.Step.Combustion qualified as Combustion
+import Swarm.Game.Step.Util
 import Swarm.Game.Universe
 import Swarm.Game.Value
 import Swarm.Game.World qualified as W
@@ -94,8 +95,6 @@ import Swarm.Language.Value
 import Swarm.Util hiding (both)
 import Swarm.Util.Effect (throwToMaybe)
 import System.Clock (TimeSpec)
-import System.Clock qualified
-import System.Random (UniformRange, uniformR)
 import Witch (From (from), into)
 import Prelude hiding (Applicative (..), lookup)
 
@@ -372,9 +371,6 @@ evalPT ::
   m Value
 evalPT t = evaluateCESK (initMachine t empty emptyStore)
 
-getNow :: Has (Lift IO) sig m => m TimeSpec
-getNow = sendIO $ System.Clock.getTime System.Clock.Monotonic
-
 -- | Create a special robot to check some hypothetical, for example the win condition.
 --
 -- Use ID (-1) so it won't conflict with any robots currently in the robot map.
@@ -416,91 +412,6 @@ runCESK (Up exn _ []) = throwError exn
 runCESK cesk = case finalValue cesk of
   Just (v, _) -> return v
   Nothing -> stepCESK cesk >>= runCESK
-
-------------------------------------------------------------
--- Some utility functions
-------------------------------------------------------------
-
--- | Set a flag telling the UI that the world needs to be redrawn.
-flagRedraw :: (Has (State GameState) sig m) => m ()
-flagRedraw = needsRedraw .= True
-
--- | Perform an action requiring a 'W.World' state component in a
---   larger context with a 'GameState'.
-zoomWorld ::
-  (Has (State GameState) sig m) =>
-  SubworldName ->
-  StateC (W.World Int Entity) Identity b ->
-  m (Maybe b)
-zoomWorld swName n = do
-  mw <- use multiWorld
-  forM (M.lookup swName mw) $ \w -> do
-    let (w', a) = run (runState w n)
-    multiWorld %= M.insert swName w'
-    return a
-
--- | Get the entity (if any) at a given location.
-entityAt :: (Has (State GameState) sig m) => Cosmic Location -> m (Maybe Entity)
-entityAt (Cosmic subworldName loc) =
-  join <$> zoomWorld subworldName (W.lookupEntityM @Int (W.locToCoords loc))
-
--- | Modify the entity (if any) at a given location.
-updateEntityAt ::
-  (Has (State GameState) sig m) =>
-  Cosmic Location ->
-  (Maybe Entity -> Maybe Entity) ->
-  m ()
-updateEntityAt cLoc@(Cosmic subworldName loc) upd = do
-  didChange <-
-    fmap (fromMaybe False) $
-      zoomWorld subworldName $
-        W.updateM @Int (W.locToCoords loc) upd
-  when didChange $
-    wakeWatchingRobots cLoc
-
--- | Get the robot with a given ID.
-robotWithID :: (Has (State GameState) sig m) => RID -> m (Maybe Robot)
-robotWithID rid = use (robotMap . at rid)
-
--- | Get the robot with a given name.
-robotWithName :: (Has (State GameState) sig m) => Text -> m (Maybe Robot)
-robotWithName rname = use (robotMap . to IM.elems . to (find $ \r -> r ^. robotName == rname))
-
--- | Generate a uniformly random number using the random generator in
---   the game state.
-uniform :: (Has (State GameState) sig m, UniformRange a) => (a, a) -> m a
-uniform bnds = do
-  rand <- use randGen
-  let (n, g) = uniformR bnds rand
-  randGen .= g
-  return n
-
--- | Given a weighting function and a list of values, choose one of
---   the values randomly (using the random generator in the game
---   state), with the probability of each being proportional to its
---   weight.  Return @Nothing@ if the list is empty.
-weightedChoice :: Has (State GameState) sig m => (a -> Integer) -> [a] -> m (Maybe a)
-weightedChoice weight as = do
-  r <- uniform (0, total - 1)
-  return $ go r as
- where
-  total = sum (map weight as)
-
-  go _ [] = Nothing
-  go !k (x : xs)
-    | k < w = Just x
-    | otherwise = go (k - w) xs
-   where
-    w = weight x
-
--- | Generate a random robot name in the form adjective_name.
-randomName :: Has (State GameState) sig m => m Text
-randomName = do
-  adjs <- use @GameState adjList
-  names <- use @GameState nameList
-  i <- uniform (bounds adjs)
-  j <- uniform (bounds names)
-  return $ T.concat [adjs ! i, "_", names ! j]
 
 ------------------------------------------------------------
 -- Debugging
@@ -564,27 +475,6 @@ ensureCanExecute c =
       let hasCaps = cap `S.member` robotCaps
       (isPrivileged || hasCaps)
         `holdsOr` Incapable FixByEquip (R.singletonCap cap) (TConst c)
-
--- | Test whether the current robot has a given capability (either
---   because it has a device which gives it that capability, or it is a
---   system robot, or we are in creative mode).
-hasCapability :: (Has (State Robot) sig m, Has (State GameState) sig m) => Capability -> m Bool
-hasCapability cap = do
-  isPrivileged <- isPrivilegedBot
-  caps <- use robotCapabilities
-  return (isPrivileged || cap `S.member` caps)
-
--- | Ensure that either a robot has a given capability, OR we are in creative
---   mode.
-hasCapabilityFor ::
-  (Has (State Robot) sig m, Has (State GameState) sig m, Has (Throw Exn) sig m) => Capability -> Term -> m ()
-hasCapabilityFor cap term = do
-  h <- hasCapability cap
-  h `holdsOr` Incapable FixByEquip (R.singletonCap cap) term
-
--- | Create an exception about a command failing.
-cmdExn :: Const -> [Text] -> Exn
-cmdExn c parts = CmdFailed c (T.unwords parts) Nothing
 
 -- | Create an exception about a command failing, with an achievement
 cmdExnWithAchievement :: Const -> [Text] -> GameplayAchievement -> Exn
@@ -1065,161 +955,6 @@ seedProgram minTime randTime thing =
     selfdestruct
   |]
 
--- Triggers the ignition of the entity underfoot with some delay.
-ignitionProgram :: Integer -> ProcessedTerm
-ignitionProgram waitTime =
-  [tmQ|
-    wait $int:waitTime;
-    try {
-      ignite down;
-      noop;
-    } {};
-    selfdestruct
-  |]
-
--- | A system program for a "combustion robot", to burn an entity
---   after it is ignited.
---
--- For efficiency, we determine a priori (i.e. the instant
--- the combustion robot is spawned) whether any neighbors will eventually
--- be burned, based on probabilities.
---
--- Note that it is possible that new neighbors may be introduced while
--- combustion is in progress. Although it may be more realistic to subject
--- these to possible combustion as well, we do not bother.
---
--- Though if we did actually want to do that, some options are:
---
--- 1. Create sub-partitions (of say, 10-tick duration) of the combustion duration
---    to re-evaluate opportunities to light adjacent entities on fire.
--- 2. Use the `watch` command to observe for changes to adjacent entities.
---    Note that if we "wake" from our `wait` due to the `watch` being triggered,
---    we would need to maintain bookkeeping of how much time is left.
--- 3. Spawn more robots whose sole purpose is to observe for changes to neighbor
---    cells. This would avoid polluting the logic of the currently burning cell
---    with logic to manage probabilities of combustion propagation.
-combustionProgram :: Integer -> Combustibility -> ProcessedTerm
-combustionProgram combustionDuration (Combustibility _ _ maybeCombustionProduct) =
-  [tmQ|
-    wait $int:combustionDuration;
-    if ($int:invQuantity > 0) {
-      try {
-        place $str:combustionProduct;
-      } {};
-    } {};
-    selfdestruct
-  |]
- where
-  (invQuantity, combustionProduct) = case maybeCombustionProduct of
-    Nothing -> (0, "")
-    Just p -> (1, p)
-
--- | We treat the 'ignition' field in the 'Combustion' record
--- as a /rate/ in a poisson distribution.
--- Ignition of neighbors depends on that particular neighbor entity's
--- combustion /rate/, but also on the duration
--- that the current entity will burn.
-igniteNeighbor ::
-  Has (State GameState) sig m =>
-  TimeSpec ->
-  Integer ->
-  Cosmic Location ->
-  m ()
-igniteNeighbor creationTime sourceDuration loc = do
-  maybeEnt <- entityAt loc
-  forM_ maybeEnt igniteEntity
- where
-  igniteEntity e =
-    when (e `hasProperty` Combustible) $ do
-      threshold <- uniform (0, 1)
-      when (probabilityOfIgnition >= threshold) $ do
-        ignitionDelayRand <- uniform (0, 1)
-        let ignitionDelay =
-              floor
-                . min (fromIntegral sourceDuration)
-                . negate
-                $ log ignitionDelayRand / rate
-        addIgnitionBot ignitionDelay e creationTime loc
-   where
-    neighborCombustibility = (e ^. entityCombustion) ? defaultCombustibility
-    rate = E.ignition neighborCombustibility
-    probabilityOfIgnition = 1 - exp (negate $ rate * fromIntegral sourceDuration)
-
--- | Construct an invisible "ignition robot" and add it to the world.
---   Its sole purpose is to delay the `ignite` command for a neighbor
---   that has been a priori determined that it shall be ignited.
-addIgnitionBot ::
-  Has (State GameState) sig m =>
-  Integer ->
-  Entity ->
-  TimeSpec ->
-  Cosmic Location ->
-  m ()
-addIgnitionBot ignitionDelay inputEntity ts loc =
-  void $
-    addTRobot $
-      mkRobot
-        ()
-        Nothing
-        "firestarter"
-        (Markdown.fromText $ T.unwords ["Delayed ignition of", (inputEntity ^. entityName) <> "."])
-        (Just loc)
-        zero
-        ( defaultEntityDisplay '*'
-            & invisible .~ True
-        )
-        (initMachine (ignitionProgram ignitionDelay) empty emptyStore)
-        []
-        []
-        True
-        False
-        ts
-
--- | Construct a "combustion robot" from entity and position
---   and add it to the world.
---   It has low priority and will be covered
---   by placed entities.
---   The "combustion bot" represents the burning of a single
---   entity; propagating the fire to neighbors is handled upstream,
---   within the `ignite` command.
-addCombustionBot ::
-  Has (State GameState) sig m =>
-  Entity ->
-  Combustibility ->
-  TimeSpec ->
-  Cosmic Location ->
-  m Integer
-addCombustionBot inputEntity combustibility ts loc = do
-  botInventory <- case maybeCombustionProduct of
-    Nothing -> return []
-    Just n -> do
-      maybeE <- uses entityMap (lookupEntityName n)
-      return $ maybe [] (pure . (1,)) maybeE
-  combustionDurationRand <- uniform durationRange
-  let combustionProg = combustionProgram combustionDurationRand combustibility
-  void $
-    addTRobot $
-      mkRobot
-        ()
-        Nothing
-        "fire"
-        (Markdown.fromText $ T.unwords ["A burning", (inputEntity ^. entityName) <> "."])
-        (Just loc)
-        zero
-        ( defaultEntityDisplay '*'
-            & displayAttr .~ AWorld "fire"
-            & displayPriority .~ 0
-        )
-        (initMachine combustionProg empty emptyStore)
-        []
-        botInventory
-        True
-        False
-        ts
-  return combustionDurationRand
- where
-  Combustibility _ durationRange maybeCombustionProduct = combustibility
-
 -- | Construct a "seed robot" from entity, time range and position,
 --   and add it to the world.  It has low priority and will be covered
 --   by placed entities.
@@ -1250,12 +985,6 @@ addSeedBot e (minT, maxT) loc ts =
         True
         False
         ts
-
--- | All functions that are used for robot step can access 'GameState' and the current 'Robot'.
---
--- They can also throw exception of our custom type, which is handled elsewhere.
--- Because of that the constraint is only 'Throw', but not 'Catch'/'Error'.
-type HasRobotStepState sig m = (Has (State GameState) sig m, Has (State Robot) sig m, Has (Throw Exn) sig m)
 
 -- | Interpret the execution (or evaluation) of a constant application
 --   to some values.
@@ -1387,30 +1116,7 @@ execConst c vs s k = do
     Harvest -> doGrab Harvest'
     Ignite -> case vs of
       [VDir d] -> do
-        let verb = "ignite"
-            verbed = "ignited"
-
-        (loc, me) <- lookInDirection d
-        -- Ensure there is an entity here.
-        e <-
-          me `isJustOrFail` ["There is nothing here to", verb <> "."]
-
-        -- Ensure it can be ignited.
-        (e `hasProperty` Combustible)
-          `holdsOrFail` ["The", e ^. entityName, "here can't be", verbed <> "."]
-
-        -- Remove the entity from the world.
-        updateEntityAt loc (const Nothing)
-        flagRedraw
-
-        -- Start burning process
-        let selfCombustibility = (e ^. entityCombustion) ? defaultCombustibility
-        createdAt <- getNow
-        combustionDurationRand <- addCombustionBot e selfCombustibility createdAt loc
-
-        let neighborLocs = map (offsetBy loc . flip applyTurn north . DRelative . DPlanar) listEnums
-        forM_ neighborLocs $ igniteNeighbor createdAt combustionDurationRand
-
+        Combustion.igniteCommand c d
         return $ Out VUnit s k
       _ -> badConst
     Swap -> case vs of
@@ -2463,19 +2169,6 @@ execConst c vs s k = do
    where
     remTime = r ^. recipeTime
 
-  deriveHeading :: HasRobotStepState sig m => Direction -> m Heading
-  deriveHeading d = do
-    orient <- use robotOrientation
-    when (isCardinal d) $ hasCapabilityFor COrient $ TDir d
-    return $ applyTurn d $ orient ? zero
-
-  lookInDirection :: HasRobotStepState sig m => Direction -> m (Cosmic Location, Maybe Entity)
-  lookInDirection d = do
-    newHeading <- deriveHeading d
-    loc <- use robotLocation
-    let nextLoc = loc `offsetBy` newHeading
-    (nextLoc,) <$> entityAt nextLoc
-
   ensureEquipped :: HasRobotStepState sig m => Text -> m Entity
   ensureEquipped itemName = do
     inst <- use equippedDevices
@@ -2730,7 +2423,7 @@ execConst c vs s k = do
         return other
 
   holdsOrFail :: (Has (Throw Exn) sig m) => Bool -> [Text] -> m ()
-  holdsOrFail a ts = a `holdsOr` cmdExn c ts
+  holdsOrFail = holdsOrFail' c
 
   holdsOrFailWithAchievement :: (Has (Throw Exn) sig m) => Bool -> [Text] -> Maybe GameplayAchievement -> m ()
   holdsOrFailWithAchievement a ts mAch = case mAch of
@@ -2738,7 +2431,7 @@ execConst c vs s k = do
     Just ach -> a `holdsOr` cmdExnWithAchievement c ts ach
 
   isJustOrFail :: (Has (Throw Exn) sig m) => Maybe a -> [Text] -> m a
-  isJustOrFail a ts = a `isJustOr` cmdExn c ts
+  isJustOrFail = isJustOrFail' c
 
   returnEvalCmp = case vs of
     [v1, v2] -> (\b -> Out (VBool b) s k) <$> evalCmp c v1 v2
@@ -2839,11 +2532,6 @@ purgeFarAwayWatches = do
 ------------------------------------------------------------
 -- Some utility functions
 ------------------------------------------------------------
-
--- | Exempts the robot from various command constraints
--- when it is either a system robot or playing in creative mode
-isPrivilegedBot :: (Has (State GameState) sig m, Has (State Robot) sig m) => m Bool
-isPrivilegedBot = (||) <$> use systemRobot <*> use creativeMode
 
 -- | Requires that the target location is within one cell.
 -- Requirement is waived if the bot is privileged.
